+++ Esp.Net.Examples.ComplexModel/Model/Entities/ReferenceData/IReferenceDataTask.cs
@@ -2,7 +2,7 @@ using System;
 
 namespace Esp.Net.Examples.ComplexModel.Model.Entities.ReferenceData
 {
-    internal interface IReferenceDataTask
+    public interface IReferenceDataTask
     {
         void BeginGetReferenceDataForCurrencyPair(Guid modelId, string currencyPair);
     }
