+++ Esp.Net.Examples.ComplexModel/Model/Events/CurrencyPairChangedEvent.cs
@@ -1,0 +1,12 @@
+﻿namespace Esp.Net.Examples.ComplexModel.Model.Events
+{
+    public class CurrencyPairChangedEvent
+    {
+        public CurrencyPairChangedEvent(string currencyPair)
+        {
+            CurrencyPair = currencyPair;
+        }
+
+        public string CurrencyPair { get; private set; }
+    }
+}
